system/requirements.txt
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
protobuf==4.24.3
requests>=2.27.1 